# swagger.yaml
openapi: 3.0.3
info:
  title: MeetFlow API
  version: 1.0.0
  description: |
    ## MeetFlow Backend REST API

    This API powers the MeetFlow SaaS platform. It is divided into four
    top-level sections:

    | Section          | Base Path          | Description                                    |
    |------------------|--------------------|------------------------------------------------|
    | **Common**       | `/api/v1/auth` & `/api/v1/account`          | Authentication & account management            |
    | **Master**       | `/api/v1/master`   | MeetFlow internal admin panel (master context) |
    | **Organization** | `/api/v1/organization` | Organisation-level management (coming soon) |
    | **Workspace**    | `/api/v1/workspace`    | Workspace & channel management (coming soon)|

    ### Authentication
    All **Master**, **Organization**, and **Workspace** routes require a valid
    JWT `accessToken` in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```
    Use the **Login** endpoint to obtain tokens.

    ### Standard Response Envelope
    Every response (success or error) is wrapped in:
    ```json
    {
      "status": {
        "response_code": 200,
        "response_message": "Success!"
      },
      "message": "Human-readable message",
      "data": { ... }
    }
    ```
    Error responses replace `data` with an `error` object containing `details`.

  contact:
    name: MeetFlow Engineering
  license:
    name: ISC
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token returned by POST /auth/login.
        Valid for 1 day (or 30 days when rememberMe is true).
...
  schemas:
...
    StatusBlock:
      type: object
      properties:
        response_code:
          type: integer
          example: 200
        response_message:
          type: string
          example: "Success!"
    SuccessEnvelope:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/StatusBlock"
        message:
          type: string
          example: "Operation successful"
        data:
          description: Payload — varies per endpoint
          nullable: true
    ErrorEnvelope:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/StatusBlock"
        message:
          type: string
          example: "Something went wrong!"
        error:
          type: object
          properties:
            details:
              type: string
              example: "Invalid credentials."
    DeletedResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              nullable: true
              example: null

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    PlanLimit:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        plan_id:
          type: integer
          format: int64
          example: 2
        limit_key:
          type: string
          example: max_users
          description: >
            Limit identifier key. Common values:
            max_users, max_workspaces, max_channels, max_meetings_per_month.
        limit_value:
          type: integer
          format: int64
          example: 50
          description: Set to 9999 to represent "Unlimited" in the UI.
        createdAt:
          type: string
          format: date-time
    AddLimitRequest:
      type: object
      required:
        - limit_key
        - limit_value
      properties:
        limit_key:
          type: string
          example: max_users
          description: >
            Snake_case limit identifier. E.g. max_users,
            max_workspaces, max_channels, max_meetings_per_month.
        limit_value:
          type: integer
          format: int64
          minimum: 0
          example: 50
          description: Set to 9999 to represent "Unlimited".
    UpdateLimitRequest:
      type: object
      required:
        - limit_value
      properties:
        limit_value:
          type: integer
          format: int64
          minimum: 0
          example: 100
          description: Set to 9999 to represent "Unlimited".

    # ── Organization ──────────────────────────────────────────────────────────
...
    BulkAddPlanLimitItem:
      type: object
      required:
        - limit_key
        - limit_value
      properties:
        limit_key:
          type: string
          maxLength: 100
          example: max_users
        limit_value:
          type: integer
          format: int64
          minimum: 0
          example: 50
    BulkAddPlanLimitRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BulkAddPlanLimitItem"
    BulkDeletePlanLimitRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: integer
            format: int64
          example: [1, 2, 3]

    # ── Bulk Organizations ────────────────────────────────────────────────────
...
  responses:
    UnauthorizedError:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            status:
              response_code: 401
              response_message: "Something went wrong!"
            message: "Access token is required."
    ForbiddenError:
      description: Authenticated but insufficient permissions (wrong context)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            status:
              response_code: 403
              response_message: "Something went wrong!"
            message: "This route is only accessible to master users."
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequestError:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
paths:
...
  /master/plans/get-plan-limits/{planId}:
    get:
      tags: [Master / Plan Limits]
      summary: Get limits for a plan
      operationId: getLimitsByPlan
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
      responses:
        "200":
          description: Plan limits retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PlanLimit"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/plans/add-plan-limit/{planId}:
    post:
      tags: [Master / Plan Limits]
      summary: Add a limit to a plan
      description: |
        Common limit_key values: max_users, max_workspaces, max_channels, max_meetings_per_month.
        Set limit_value to 9999 to represent "Unlimited" in the UI.
      operationId: addLimit
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLimitRequest"
            example:
              limit_key: max_users
              limit_value: 50
      responses:
        "201":
          description: Limit added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PlanLimit"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/plans/update-plan-limit/{planId}/{limitId}:
    patch:
      tags: [Master / Plan Limits]
      summary: Update a plan limit
      operationId: updateLimit
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
        - name: limitId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLimitRequest"
            example:
              limit_value: 100
      responses:
        "200":
          description: Limit updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PlanLimit"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/plans/delete-plan-limit/{planId}/{limitId}:
    delete:
      tags: [Master / Plan Limits]
      summary: Delete a plan limit
      operationId: deleteLimit
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
        - name: limitId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 5
      responses:
        "200":
          description: Limit deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ─── Bulk Plan Limit Operations ────────────────────────────────────────────
  /master/plans/bulk-add-limits/{planId}:
    post:
      tags: [Master / Plan Limits]
      summary: Bulk add limits to a plan
      description: |
        Add multiple quantitative limits to a plan in a single request.
        This is the recommended way to seed all limits for a plan at once.

        **Common limit_key values:**
        `max_members`, `max_workspaces`, `max_channels_per_workspace`,
        `max_meetings_per_month`, `max_meeting_duration_minutes`,
        `max_meeting_participants`, `max_file_storage_mb`,
        `message_history_days`, `max_guests`, `max_ai_queries_per_month`.

        **Limits:** Maximum 100 items. Each limit_key must be unique in the request.
      operationId: bulkAddLimits
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          description: ID of the plan
          schema:
            type: integer
            format: int64
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAddPlanLimitRequest"
            example:
              items:
                - limit_key: max_members
                  limit_value: 100
                - limit_key: max_workspaces
                  limit_value: 10
                - limit_key: max_meetings_per_month
                  limit_value: 200
                - limit_key: message_history_days
                  limit_value: -1
      responses:
        "200":
          description: Bulk add limits completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          successful:
                            type: array
                            items:
                              $ref: "#/components/schemas/PlanLimit"
                          failed:
                            type: array
                          summary:
                            type: object
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/plans/bulk-delete-limits/{planId}:
    delete:
      tags: [Master / Plan Limits]
      summary: Bulk delete limits from a plan
      description: |
        Delete multiple limits from a plan by limit IDs.
        **Limits:** Maximum 100 limit IDs per request.
      operationId: bulkDeleteLimits
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          description: ID of the plan
          schema:
            type: integer
            format: int64
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeletePlanLimitRequest"
            example:
              ids: [1, 2, 3]
      responses:
        "200":
          description: Bulk delete limits completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ═══════════════════════════════════════════════════════════════════════════
  # MASTER — Features
  # ═══════════════════════════════════════════════════════════════════════════
...